# test/api/env/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IIA Server API",
    "description": "IIA (Intelligent Information Assistant) Server API Documentation",
    "contact": {
      "name": "Charles",
      "email": "inforassistant@foxmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9424",
      "description": "Local Development Environment"
    },
    {
      "url": "http://10.15.0.21:9424",
      "description": "Test Environment"
    }
  ],
  "security": [
    {
      "bearer-jwt": []
    }
  ],
...
  "paths": {
...
    "/api/reminder/task/batch-update-position": {
      "post": {
        "tags": [
          "task-controller"
        ],
        "operationId": "batchUpdatePosition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdatePositionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reminder/project/update": {
      "post": {
        "tags": [
          "Project Management"
        ],
        "summary": "Update Project",
        "description": "Update project information including name, description, color, and icon",
        "operationId": "updateById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Project updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsProject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsProject"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsProject"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsProject"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
    "/api/reminder/project/get": {
      "post": {
        "tags": [
          "Project Management"
        ],
        "summary": "Query Projects",
        "description": "Query projects by archived flag and isAll flag",
        "operationId": "get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectGetDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Projects queried successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsListProject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsListProject"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsListProject"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
    "/api/reminder/project/delete": {
      "post": {
        "tags": [
          "Project Management"
        ],
        "summary": "Delete Project",
        "description": "Delete a project with options to keep tasks or move them to another project",
        "operationId": "delete_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDeleteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Project deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
    "/api/reminder/project/create": {
      "post": {
        "tags": [
          "Project Management"
        ],
        "summary": "Create Project",
        "description": "Create a new reminder project with name, description, color, and icon",
        "operationId": "create_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsProject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsProject"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsProject"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
    "/api/reminder/project/batch-update-position": {
      "post": {
        "tags": [
          "Project Management"
        ],
        "summary": "Batch Update Project Positions",
        "description": "Update positions of multiple projects in batch",
        "operationId": "batchUpdatePosition_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdatePositionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Project positions updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsListPosition"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsListPosition"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtilsListPosition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseUtilsVoid": {
        "type": "object",
        "description": "Standard API response structure",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "Response status code",
            "example": 200,
            "maximum": 599,
            "minimum": 100
          },
          "msg": {
            "type": "string",
            "description": "Response message",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "Response data payload, can be null for error responses"
          }
        }
      },
...
      "BatchUpdatePositionDTO": {
        "type": "object",
        "description": "Batch update position request data",
        "properties": {
          "pos": {
            "type": "array",
            "description": "Position list",
            "example": [
              {
                "itemId": 1,
                "sortOrder": 0
              },
              {
                "itemId": 2,
                "sortOrder": 1
              }
            ],
            "items": {
              "$ref": "#/components/schemas/Position"
            }
          }
        },
        "required": [
          "pos"
        ]
      },
      "Position": {
        "type": "object",
        "description": "Position information for sorting items",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64",
            "description": "Item ID",
            "example": 1
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "description": "Sort order",
            "example": 0
          }
        },
        "required": [
          "itemId",
          "sortOrder"
        ]
      },
...
      "ProjectUpdateDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          },
          "color": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "projectId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "name"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "isArchived": {
            "type": "boolean"
          },
          "operationId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResponseUtilsProject": {
        "type": "object",
        "description": "Standard API response structure",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "Response status code",
            "example": 200,
            "maximum": 599,
            "minimum": 100
          },
          "msg": {
            "type": "string",
            "description": "Response message",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/Project",
            "description": "Response data payload, can be null for error responses"
          }
        }
      },
      "ProjectGetDTO": {
        "type": "object",
        "description": "Project query request data",
        "properties": {
          "archived": {
            "type": "boolean",
            "description": "Whether to query archived projects",
            "example": false
          },
          "isAll": {
            "type": "boolean",
            "description": "Whether to return both archived and unarchived projects",
            "example": false
          }
        },
        "required": [
          "archived"
        ]
      },
      "ResponseUtilsListProject": {
        "type": "object",
        "description": "Standard API response structure",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "Response status code",
            "example": 200,
            "maximum": 599,
            "minimum": 100
          },
          "msg": {
            "type": "string",
            "description": "Response message",
            "example": "success"
          },
          "data": {
            "type": "array",
            "description": "Response data payload, can be null for error responses",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          }
        }
      },
      "ProjectDeleteDTO": {
        "type": "object",
        "description": "Project delete request data",
        "properties": {
          "projectId": {
            "type": "integer",
            "format": "int64",
            "description": "Project ID",
            "example": 1
          },
          "keepTasks": {
            "type": "boolean",
            "description": "Keep tasks when deleting the project",
            "example": true
          },
          "targetProject": {
            "type": "boolean",
            "description": "Delete the target project",
            "example": false
          },
          "targetProjectId": {
            "type": "integer",
            "format": "int64",
            "description": "Target project ID"
          }
        },
        "required": [
          "keepTasks",
          "projectId",
          "targetProject",
          "targetProjectId"
        ]
      },
      "ProjectCreateDTO": {
        "type": "object",
        "description": "Project create request data",
        "properties": {
          "name": {
            "type": "string",
            "description": "Project name",
            "example": "P1",
            "maxLength": 20,
            "minLength": 1
          },
          "description": {
            "type": "string",
            "description": "Project description",
            "example": "Project 1",
            "maxLength": 500,
            "minLength": 1
          },
          "color": {
            "type": "string",
            "description": "Project color",
            "example": "#409EFF"
          },
          "icon": {
            "type": "string",
            "description": "Project icon",
            "example": "icon-1"
          }
        },
        "required": [
          "color",
          "description",
          "icon",
          "name"
        ]
      },
      "ResponseUtilsListPosition": {
        "type": "object",
        "description": "Standard API response structure",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "Response status code",
            "example": 200,
            "maximum": 599,
            "minimum": 100
          },
          "msg": {
            "type": "string",
            "description": "Response message",
            "example": "success"
          },
          "data": {
            "type": "array",
            "description": "Response data payload, can be null for error responses",
            "items": {
              "$ref": "#/components/schemas/Position"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "type": "http",
        "description": "JWT authentication token, format: Bearer {token}",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "x-apifox": {
          "variable": "access_token"
        }
      }
    }
  }
}